+++ include/xpln/obj/attributes/AttrBlend.h
@@ -34,6 +34,8 @@
 
 namespace xobj {
 
+class AbstractWriter;
+
 /**************************************************************************************************/
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 /**************************************************************************************************/
@@ -95,6 +97,16 @@ public:
     XpObjLib void setRatio(float ratio);
     XpObjLib float ratio() const;
     XpObjLib eType type() const;
+
+    //-------------------------------------------------------------------------
+
+    /*!
+     * \note For internal use only.
+     * \return String with default values for simulator.
+     *         It is needed when attribute has been enabled before
+     *         and now should be disabled.
+     */
+    XpObjLib static std::string objDisableStr();
 
     //-------------------------------------------------------------------------